+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.3
 + Fix edge case where data is not loaded when dynamic imports are invoked while changing locales.
++ Automatically embed external locales using u27n manifests.
 
 ## 0.3.2
 + Add noop environment module in case code is loaded in a non-webpack runtime.